refactor: home page, add the sections for other types of movies
