+++ src/app/page.js
@@ -2,27 +2,19 @@
 import React, { useEffect, useRef, useState } from "react";
 
 import HomeSlider from "./components/home-slider";
-import GlobalApi from "./services/globalApi";
 import Hero from "./components/hero";
 
 export default function Home() {
-  const [movieList, setMovieList] = useState([]);
  
-  useEffect(() => {
-    getPopMovies();
-  }, []);
-
-  const getPopMovies = () => {
-    GlobalApi.getTrendingVideos.then(resp=>{
-      setMovieList(resp.data.results)
-  })
-  };
 
   return (
     <>
 
       <Hero/>
-      <HomeSlider movieList={movieList} title={"Latest Movies"} background={"gradient-blue"}/>
+      <HomeSlider type={"top_rated"} title={"Top Rated"} background={"gradient-blue"}/>
+      <HomeSlider type={"popular"} title={"Popular"} background={"gradient-blue"}/>
+      <HomeSlider type={"now_playing"} title={"Now Playing"} background={"gradient-blue"}/>
+      <HomeSlider type={"upcoming"} title={"Upcoming"} background={"gradient-blue"}/>
       {/* <div>test</div> */}
      
     </>
